Fix GHSA fix with respect to int64 update in #2415 (#2431)

+++ internal/controlplane/handlers_githubwebhooks.go
@@ -326,7 +326,7 @@ func (s *Server) registerWebhookForRepository(
 	regResult.Repository = &pb.Repository{
 		Name:          repoGet.GetName(),
 		Owner:         repoGet.GetOwner().GetLogin(),
-		RepoId:        int32(repoGet.GetID()),
+		RepoId:        repoGet.GetID(),
 		HookId:        mhook.GetID(),
 		HookUrl:       mhook.GetURL(),
 		DeployUrl:     repoGet.GetDeploymentsURL(),
